キャスト関係のwarningを消した
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@9017 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Physics/PHHapticEngine.cpp
@@ -180,7 +180,7 @@ bool PHShapePairForHaptic::CompIntermediateRepresentation(Posed curShapePoseW[2]
 		ir->interpolation_pose = curShapePoseW[0];
 		irs.push_back(ir);
 	}
-	nIrsNormal = irs.size();
+	nIrsNormal = (int)irs.size();
 	return true;
 }
